compatibility fixes for Solaris 8/gcc 2.95
Original commit message from CVS:
compatibility fixes for Solaris 8/gcc 2.95
* configure.ac:
include libintl libs in LDFLAGS
* gstvalue.c (gst_value_deserialize_buffer):
cast isxdigit stuff to int to silence compiler warning

+++ gst/gstvalue.c
@@ -889,7 +889,7 @@ gst_value_deserialize_buffer (GValue * dest, const char *s)
   buffer = gst_buffer_new_and_alloc (len / 2);
   data = GST_BUFFER_DATA (buffer);
   for (i = 0; i < len / 2; i++) {
-    if (!isxdigit (s[i * 2]) || !isxdigit (s[i * 2 + 1])) {
+    if (!isxdigit ((int) s[i * 2]) || !isxdigit ((int) s[i * 2 + 1])) {
       ret = FALSE;
       break;
     }
